ekf_control: move controlMagFusion and related functions to mag_control.cpp
