+++ EKF/control.cpp
@@ -1304,300 +1304,6 @@ void Ekf::controlDragFusion()
 	}
 }
 
-void Ekf::controlMagFusion()
-{
-	// If we are on ground, store the local position and time to use as a reference
-	// Also reset the flight alignment flag so that the mag fields will be re-initialised next time we achieve flight altitude
-	if (!_control_status.flags.in_air) {
-		_last_on_ground_posD = _state.pos(2);
-		_control_status.flags.mag_aligned_in_flight = false;
-		_num_bad_flight_yaw_events = 0;
-	}
-
-	if ((_params.mag_fusion_type >= MAG_FUSE_TYPE_NONE)
-	    || _control_status.flags.mag_fault
-	    || !_control_status.flags.tilt_align) {
-		stopMagFusion();
-		return;
-	}
-
-	if (canRunMagFusion()) {
-		if (_control_status.flags.in_air) {
-			checkHaglYawResetReq();
-			runInAirYawReset();
-			runVelPosReset();
-
-		} else {
-			runOnGroundYawReset();
-		}
-
-		// Determine if we should use simple magnetic heading fusion which works better when
-		// there are large external disturbances or the more accurate 3-axis fusion
-		switch (_params.mag_fusion_type) {
-		case MAG_FUSE_TYPE_AUTO:
-			selectMagAuto();
-			break;
-
-		case MAG_FUSE_TYPE_AUTOFW:
-			selectMagAuto();
-			controlMagStateOnlyFusion();
-			break;
-
-		case MAG_FUSE_TYPE_HEADING:
-		case MAG_FUSE_TYPE_INDOOR:
-			startMagHdgFusion();
-			break;
-
-		case MAG_FUSE_TYPE_3D:
-			startMag3DFusion();
-			break;
-
-		default:
-			stopMagFusion();
-			break;
-		}
-
-		checkMagDeclRequired();
-		checkMagInhibition();
-
-		if (_control_status.flags.yaw_align) {
-			runMagAndMagDeclFusions();
-		}
-	}
-}
-
-bool Ekf::canRunMagFusion() const
-{
-	// check for new magnetometer data that has fallen behind the fusion time horizon
-	// If we are using external vision data or GPS-heading for heading then no magnetometer fusion is used
-	return !_control_status.flags.ev_yaw && !_control_status.flags.gps_yaw && _mag_data_ready;
-}
-
-void Ekf::checkHaglYawResetReq()
-{
-	// We need to reset the yaw angle after climbing away from the ground to enable
-	// recovery from ground level magnetic interference.
-	if (!_control_status.flags.mag_aligned_in_flight) {
-		// Check if height has increased sufficiently to be away from ground magnetic anomalies
-		// and request a yaw reset if not already requested.
-		static constexpr float mag_anomalies_max_hagl = 1.5f;
-		const bool above_mag_anomalies = (getTerrainVPos() - _state.pos(2)) > mag_anomalies_max_hagl;
-		_mag_yaw_reset_req = _mag_yaw_reset_req || above_mag_anomalies;
-	}
-}
-
-float Ekf::getTerrainVPos() const
-{
-	return isTerrainEstimateValid() ? _terrain_vpos : _last_on_ground_posD;
-}
-
-void Ekf::runOnGroundYawReset()
-{
-	if (_mag_yaw_reset_req && isYawResetAuthorized()) {
-		const bool has_realigned_yaw = canResetMagHeading()
-					       ? resetMagHeading(_mag_sample_delayed.mag)
-					       : false;
-
-		_control_status.flags.yaw_align = _control_status.flags.yaw_align || has_realigned_yaw;
-		_mag_yaw_reset_req = !has_realigned_yaw;
-	}
-}
-
-void Ekf::runInAirYawReset()
-{
-	if (_mag_yaw_reset_req && isYawResetAuthorized()) {
-		const bool has_realigned_yaw = canRealignYawUsingGps()
-					       ? realignYawGPS()
-					       : canResetMagHeading()
-						      ? resetMagHeading(_mag_sample_delayed.mag)
-						      : false;
-
-		_control_status.flags.yaw_align = _control_status.flags.yaw_align || has_realigned_yaw;
-		_mag_yaw_reset_req = !has_realigned_yaw;
-		_control_status.flags.mag_aligned_in_flight = has_realigned_yaw;
-	}
-}
-
-bool Ekf::isYawResetAuthorized() const
-{
-	return !_mag_use_inhibit;
-}
-
-void Ekf::runVelPosReset()
-{
-	if (_velpos_reset_request) {
-		resetVelocity();
-		resetPosition();
-		_velpos_reset_request = false;
-	}
-}
-
-void Ekf::selectMagAuto()
-{
-	check3DMagFusionSuitability();
-	canUse3DMagFusion() ? startMag3DFusion() : startMagHdgFusion();
-}
-
-void Ekf::check3DMagFusionSuitability()
-{
-	checkYawAngleObservability();
-	checkMagBiasObservability();
-
-	if (isMagBiasObservable() || isYawAngleObservable()) {
-		_time_last_mov_3d_mag_suitable = _imu_sample_delayed.time_us;
-	}
-}
-
-void Ekf::checkYawAngleObservability()
-{
-	// Check if there has been enough change in horizontal velocity to make yaw observable
-	// Apply hysteresis to check to avoid rapid toggling
-	_yaw_angle_observable = _yaw_angle_observable
-				? _accel_lpf_NE.norm() > _params.mag_acc_gate
-				: _accel_lpf_NE.norm() > 2.0f * _params.mag_acc_gate;
-
-	_yaw_angle_observable = _yaw_angle_observable
-				&& (_control_status.flags.gps || _control_status.flags.ev_pos); // Do we have to add ev_vel here?
-}
-
-void Ekf::checkMagBiasObservability()
-{
-	// check if there is enough yaw rotation to make the mag bias states observable
-	if (!_mag_bias_observable && (fabsf(_yaw_rate_lpf_ef) > _params.mag_yaw_rate_gate)) {
-		// initial yaw motion is detected
-		_mag_bias_observable = true;
-
-	} else if (_mag_bias_observable) {
-		// require sustained yaw motion of 50% the initial yaw rate threshold
-		const float yaw_dt = 1e-6f * (float)(_imu_sample_delayed.time_us - _time_yaw_started);
-		const float min_yaw_change_req =  0.5f * _params.mag_yaw_rate_gate * yaw_dt;
-		_mag_bias_observable = fabsf(_yaw_delta_ef) > min_yaw_change_req;
-	}
-
-	_yaw_delta_ef = 0.0f;
-	_time_yaw_started = _imu_sample_delayed.time_us;
-}
-
-bool Ekf::canUse3DMagFusion() const
-{
-	// Use of 3D fusion requires an in-air heading alignment but it should not
-	// be used when the heading and mag biases are not observable for more than 2 seconds
-	return _control_status.flags.mag_aligned_in_flight
-	       && ((_imu_sample_delayed.time_us - _time_last_mov_3d_mag_suitable) < (uint64_t)2e6);
-}
-
-void Ekf::controlMagStateOnlyFusion()
-{
-	/*
-	Control switch-over between only updating the mag states to updating all states
-	When flying as a fixed wing aircraft, a misaligned magnetometer can cause an error in pitch/roll and accel bias estimates.
-	When MAG_FUSE_TYPE_AUTOFW is selected and the vehicle is flying as a fixed wing, then magnetometer fusion is only allowed
-	to access the magnetic field states.
-	*/
-	_control_status.flags.update_mag_states_only =  _control_status.flags.fixed_wing;
-
-	// For the first 5 seconds after switching to 3-axis fusion we allow the magnetic field state estimates to stabilise
-	// before they are used to constrain heading drift
-	_flt_mag_align_converging = ((_imu_sample_delayed.time_us - _flt_mag_align_start_time) < (uint64_t)5e6);
-
-	if (_control_status.flags.mag_3D
-	    && _control_status_prev.flags.update_mag_states_only
-	    && !_control_status.flags.update_mag_states_only) {
-		// When re-commencing use of magnetometer to correct vehicle states
-		// set the field state variance to the observation variance and zero
-		// the covariance terms to allow the field states re-learn rapidly
-		clearMagCov();
-
-		for (uint8_t index = 0; index <= 5; index ++) {
-			P[index + 16][index + 16] = sq(_params.mag_noise);
-		}
-
-		// save covariance data for re-use when auto-switching between heading and 3-axis fusion
-		saveMagCovData();
-	}
-}
-
-void Ekf::checkMagDeclRequired()
-{
-	// if we are using 3-axis magnetometer fusion, but without external NE aiding,
-	// then the declination must be fused as an observation to prevent long term heading drift
-	// fusing declination when gps aiding is available is optional, but recommended to prevent
-	// problem if the vehicle is static for extended periods of time
-	const bool user_selected = (_params.mag_declination_source & MASK_FUSE_DECL);
-	const bool not_using_ne_aiding = !_control_status.flags.gps;
-	_control_status.flags.mag_dec = (_control_status.flags.mag_3D && (not_using_ne_aiding || user_selected));
-}
-
-void Ekf::checkMagInhibition()
-{
-	_mag_use_inhibit = shouldInhibitMag();
-	if (!_mag_use_inhibit) { 
-		_mag_use_not_inhibit_us = _imu_sample_delayed.time_us;
-	}
-
-	// If magnetometer use has been inhibited continuously then a yaw reset is required for a valid heading
-	if (uint32_t(_imu_sample_delayed.time_us - _mag_use_not_inhibit_us) > (uint32_t)5e6) {
-		_mag_inhibit_yaw_reset_req = true;
-	}
-}
-
-bool Ekf::shouldInhibitMag() const
-{
-	// If the user has selected auto protection against indoor magnetic field errors, only use the magnetometer
-	// if a yaw angle relative to true North is required for navigation. If no GPS or other earth frame aiding
-	// is available, assume that we are operating indoors and the magnetometer should not be used.
-	// Also inhibit mag fusion when a strong magnetic field interference is detected
-	const bool user_selected = (_params.mag_fusion_type == MAG_FUSE_TYPE_INDOOR);
-	const bool not_using_gps = !(_params.fusion_mode & MASK_USE_GPS) || !_control_status.flags.gps;
-	const bool not_using_evpos = !(_params.fusion_mode & MASK_USE_EVPOS) || !_control_status.flags.ev_pos;
-	const bool not_using_evvel = !(_params.fusion_mode & MASK_USE_EVVEL) || !_control_status.flags.ev_vel;
-	const bool not_selected_evyaw =  !(_params.fusion_mode & MASK_USE_EVYAW);
-	const bool heading_not_required_for_navigation = not_using_gps &&
-							 not_using_evpos &&
-							 not_using_evvel &&
-							 not_selected_evyaw;
-
-	return (user_selected && heading_not_required_for_navigation)
-	       || isStrongMagneticDisturbance();
-}
-
-bool Ekf::isStrongMagneticDisturbance() const
-{
-	return _mag_sample_delayed.mag.length() > 2.f * _mag_strength_gps;
-}
-
-void Ekf::runMagAndMagDeclFusions()
-{
-	if (_control_status.flags.mag_3D) {
-		run3DMagAndDeclFusions();
-
-	} else if (_control_status.flags.mag_hdg) {
-		fuseHeading();
-	}
-}
-
-void Ekf::run3DMagAndDeclFusions()
-{
-	if (!_mag_decl_cov_reset) {
-		// After any magnetic field covariance reset event the earth field state
-		// covariances need to be corrected to incorporate knowedge of the declination
-		// before fusing magnetomer data to prevent rapid rotation of the earth field
-		// states for the first few observations.
-		fuseDeclination(0.02f);
-		_mag_decl_cov_reset = true;
-		fuseMag();
-
-	} else {
-		// The normal sequence is to fuse the magnetometer data first before fusing
-		// declination angle at a higher uncertainty to allow some learning of
-		// declination angle over time.
-		fuseMag();
-		if (_control_status.flags.mag_dec) {
-			fuseDeclination(0.5f);
-		}
-	}
-}
-
 void Ekf::controlVelPosFusion()
 {
 	// if we aren't doing any aiding, fake GPS measurements at the last known position to constrain drift
